Add custom hook to consume CTN layers from AGOL
